[⚙]🐛fix: SecondHandBid 사용자 처리
