+++ src/main/java/ddwu/project/mdm_ver2/domain/secondhand/controller/sh/SecondHandController.java
@@ -1,8 +1,8 @@
 package ddwu.project.mdm_ver2.domain.secondhand.controller.sh;
 
-import ddwu.project.mdm_ver2.domain.secondhand.dto.SecondHandResponse;
+import ddwu.project.mdm_ver2.domain.secondhand.dto.sh.SecondHandResponse;
 import ddwu.project.mdm_ver2.domain.secondhand.entity.SecondHand;
-import ddwu.project.mdm_ver2.domain.secondhand.dto.SecondHandRequest;
+import ddwu.project.mdm_ver2.domain.secondhand.dto.sh.SecondHandRequest;
 import ddwu.project.mdm_ver2.domain.secondhand.service.SecondHandService;
 import ddwu.project.mdm_ver2.global.exception.CustomResponse;
 import lombok.AllArgsConstructor;
