+++ src/main/java/ddwu/project/mdm_ver2/domain/secondhand/dto/sh/SecondHandRequest.java
@@ -1,18 +1,17 @@
-package ddwu.project.mdm_ver2.domain.secondhand.dto;
+package ddwu.project.mdm_ver2.domain.secondhand.dto.sh;
 
-import ddwu.project.mdm_ver2.domain.secondhand.entity.SecondHandBid;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
-import java.util.List;
+import java.sql.Date;
 
 @Data
 @Builder
 @AllArgsConstructor
 @NoArgsConstructor
-public class SecondHandResponse {
+public class SecondHandRequest {
 
     private Long userId; /* 중고거래 요청자 (가격 제안) */
     private String name;
@@ -20,7 +19,5 @@ public class SecondHandResponse {
     private int price;
     private String imgUrl;
     private String content;
-    private List<SecondHandBid> secondHandBidList;
-    private Character favState;
-    private Character userState; /* 로그인 X, 작성자와 현재 사용자 다름 ('n') */
+
 }
