+++ src/main/java/ddwu/project/mdm_ver2/domain/secondhand/dto/sh/SecondHandResponse.java
@@ -1,17 +1,18 @@
-package ddwu.project.mdm_ver2.domain.secondhand.dto;
+package ddwu.project.mdm_ver2.domain.secondhand.dto.sh;
 
+import ddwu.project.mdm_ver2.domain.secondhand.entity.SecondHandBid;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
-import java.sql.Date;
+import java.util.List;
 
 @Data
 @Builder
 @AllArgsConstructor
 @NoArgsConstructor
-public class SecondHandRequest {
+public class SecondHandResponse {
 
     private Long userId; /* 중고거래 요청자 (가격 제안) */
     private String name;
@@ -19,5 +20,7 @@ public class SecondHandRequest {
     private int price;
     private String imgUrl;
     private String content;
-
+    private List<SecondHandBid> secondHandBidList;
+    private Character favState;
+    private Character userState; /* 로그인 X, 작성자와 현재 사용자 다름 ('n') */
 }
