+++ src/main/java/ddwu/project/mdm_ver2/domain/secondhand/service/SecondHandBidService.java
@@ -1,10 +1,8 @@
 package ddwu.project.mdm_ver2.domain.secondhand.service;
 
-import ddwu.project.mdm_ver2.domain.category.entity.Category;
-import ddwu.project.mdm_ver2.domain.secondhand.dto.SecondHandRequest;
 import ddwu.project.mdm_ver2.domain.secondhand.entity.SecondHand;
 import ddwu.project.mdm_ver2.domain.secondhand.entity.SecondHandBid;
-import ddwu.project.mdm_ver2.domain.secondhand.dto.SecondHandBidRequest;
+import ddwu.project.mdm_ver2.domain.secondhand.dto.shBid.SecondHandBidRequest;
 import ddwu.project.mdm_ver2.domain.user.entity.User;
 import ddwu.project.mdm_ver2.domain.user.repository.UserRepository;
 import ddwu.project.mdm_ver2.global.exception.CustomResponse;
@@ -17,6 +15,7 @@ import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Service;
 import org.webjars.NotFoundException;
 
+import java.security.Principal;
 import java.util.List;
 
 @Service
@@ -77,25 +76,29 @@ public class SecondHandBidService {
     }
 
     /* 제안 수정 */
-    public CustomResponse<SecondHandBid> updateShBid(String userEmail, Long shBidId, SecondHandBidRequest request) {
+    public CustomResponse<SecondHandBid> updateShBid(Principal principal, Long shBidId, SecondHandBidRequest request) {
         try {
             SecondHandBid secondHandBid = shBidRepository.findById(shBidId)
                     .orElseThrow(() -> new NotFoundException("상품 요청을 찾을 수 없습니다."));
 
-            User user = userRepository.findByEmail(userEmail)
-                    .orElseThrow(() -> new NotFoundException("사용자를 찾을 수 없습니다."));
-
-            if((user.getId()) == (secondHandBid.getBidUserId())) {
-                if(secondHandBid != null) {
-                    secondHandBid.setPrice(request.getPrice());
-                    SecondHandBid updateSecondHandBid = shBidRepository.saveAndFlush(secondHandBid);
-
-                    return CustomResponse.onSuccess(updateSecondHandBid);
-                } else {
-                    return CustomResponse.onFailure(HttpStatus.NOT_FOUND.value(), "상품 요청을 찾을 수 없습니다.");
-                }
+            if(principal == null) {
+                return CustomResponse.onFailure(HttpStatus.METHOD_NOT_ALLOWED.value(), "중고 거래 상품을 수정할 수 없습니다.");
             } else {
-                return CustomResponse.onFailure(HttpStatus.METHOD_NOT_ALLOWED.value(), "요청을 수정할 수 없습니다.");
+                User user = userRepository.findByEmail(principal.getName())
+                        .orElseThrow(() -> new NotFoundException("사용자를 찾을 수 없습니다."));
+
+                if((user.getId()) == (secondHandBid.getBidUserId())) {
+                    if(secondHandBid != null) {
+                        secondHandBid.setPrice(request.getPrice());
+                        SecondHandBid updateSecondHandBid = shBidRepository.saveAndFlush(secondHandBid);
+
+                        return CustomResponse.onSuccess(updateSecondHandBid);
+                    } else {
+                        return CustomResponse.onFailure(HttpStatus.NOT_FOUND.value(), "상품 요청을 찾을 수 없습니다.");
+                    }
+                } else {
+                    return CustomResponse.onFailure(HttpStatus.METHOD_NOT_ALLOWED.value(), "요청을 수정할 수 없습니다.");
+                }
             }
         } catch (Exception e) {
             return CustomResponse.onFailure(HttpStatus.INTERNAL_SERVER_ERROR.value(), e.getMessage());
@@ -104,7 +107,7 @@ public class SecondHandBidService {
 
     /* 제안 삭제 */
     @Transactional
-    public CustomResponse<Void> deleteSecondHandBid(String userEmail, Long shId, Long shBidId) {
+    public CustomResponse<Void> deleteSecondHandBid(Principal principal, Long shId, Long shBidId) {
         try {
             SecondHand secondHand = secondHandRepository.findById(shId)
                     .orElseThrow(() -> new NotFoundException("중고거래 상품을 찾을 수 없습니다."));
@@ -112,21 +115,21 @@ public class SecondHandBidService {
             SecondHandBid secondHandBid = shBidRepository.findById(shBidId)
                     .orElseThrow(() -> new NotFoundException("상품 요청을 찾을 수 없습니다."));
 
-            User user = userRepository.findByEmail(userEmail)
-                    .orElseThrow(() -> new NotFoundException("사용자를 찾을 수 없습니다."));
-
-            if((user.getId()) == (secondHandBid.getBidUserId())) {
-                if(secondHandBid == null) {
-                    throw new ResourceNotFoundException("secondHandBid", "shBidId", shBidId);
-                }
+            if(principal == null) {
+                return CustomResponse.onFailure(HttpStatus.METHOD_NOT_ALLOWED.value(), "중고 거래 상품을 수정할 수 없습니다.");
             } else {
-                return CustomResponse.onFailure(HttpStatus.METHOD_NOT_ALLOWED.value(), "요청을 삭제할 수 없습니다.");
+                User user = userRepository.findByEmail(principal.getName())
+                        .orElseThrow(() -> new NotFoundException("사용자를 찾을 수 없습니다."));
+
+                if((user.getId()) == (secondHandBid.getBidUserId())) {
+                    setSecondHandBidCnt(secondHand, secondHand.getBidCnt(), "delete");
+                    shBidRepository.deleteById(shBidId);
+
+                    return CustomResponse.onSuccess(null);
+                } else {
+                    return CustomResponse.onFailure(HttpStatus.METHOD_NOT_ALLOWED.value(), "요청을 삭제할 수 없습니다.");
+                }
             }
-
-            setSecondHandBidCnt(secondHand, secondHand.getBidCnt(), "delete");
-            shBidRepository.deleteById(shBidId);
-
-            return CustomResponse.onSuccess(null);
         } catch (Exception e) {
             return CustomResponse.onFailure(HttpStatus.INTERNAL_SERVER_ERROR.value(), e.getMessage());
         }
